Adds Translate()
Translate() runs at the start of PEMDAS(). Its purpose is to translate the version of count that was used for the GUI, which includes html tags and other characters that cant be processed by the calculator, and translates it to a version that can be processed by the calculator.

Features:

- changes html tags to parentheses

- if + - / *  or ^ are not present inbetween two components, a multiplication symbol is added (e.g. '(2)(3)' is '(2) * (3)')

- changes '-log' to '-1 * log'

-changes π to math.pi()

- replaces double spaces with single spaces