Modifying multiplication_and_division() and addition_and_subtraction()
multiplication_and_division():

Adds a while loop (line 149) that runs while '*', and '÷' are present in count.

adds break statements to the end of each if statement so that the for loop adjusts to the new length of components

addition_and_subtraction():
Adds a while loop (line 149) that runs while ' +' , and  '- ' are present in count.

adds break statements to the end of each if statement so that the for loop adjusts to the new length of components