Updates GUI
adds find_distance(), which determines how many characters should be deleted or moved passed

This was implemented because in some cases (e.g. an html tag) if one character was deleted or moved passed, it would mess up the whole equation or reveal and nullify html tags. For example,  "<sup>" turning into "<sup_>" or "<sup", and "log" turning into "lo" or "lo_g". These would return an error if the user tried to calculate and "<sup>" would no longer be superscripting the desired characters.

Adds addFunction() and view() which allow the user to add a function to the graph and then view said graph.