Create GraphGUI
GraphGUI creates a 700 x 500 window, with the Graph placed on the left, and buttons on the right, such as the zoom buttons, which allow the user to increase or decrease the zoom of the graph.

I made this new class/window instead of just expanding the Graph and adding buttons because I wanted to update the Graph without updating the entire thing. With this implementation, I can update the Graph widget while the rest of the GraphGUI remains unaffected.