Update Graph.py
Adds the @cache decorator to get_lines(). This improves performance when adjusting the transparency by a lot.

Adds the zoom parameter to get_lines().  while zoom is not actually used, it tells the @cache decorator that the zoom has changed. This solves a problem I had where when I increased/decreased the zoom the functions would not change. This was because even though the zoom was different and it should be returning a different set of lines/points, the method being called was the same as before, so it returned the cached return value.

adds the change_opacity() method. This allows the GraphGUI to change the opacity of individual functions.